Rename package and project structure to `il.co.sysbind.moodledev`.
Renamed all occurrences of `il.co.sysbind.intellij.moodledev` to `il.co.sysbind.moodledev` across the project. Updated corresponding import statements and the plugin configuration in the plugin.xml file.
