+++ gradle.properties
@@ -1,6 +1,6 @@
 # IntelliJ Platform Artifacts Repositories -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = il.co.sysbind.intellij.moodledev
+pluginGroup = il.co.sysbind.moodledev
 pluginName = moodle-dev
 pluginRepositoryUrl = https://github.com/SysBind/moodle-dev
 # SemVer format -> https://semver.org
