--- src/main/kotlin/il/co/sysbind/intellij/moodledev/MoodleIcons.kt
@@ -1,7 +1,0 @@
-package il.co.sysbind.intellij.moodledev
-
-import com.intellij.openapi.util.IconLoader
-
-object MoodleIcons {
-    val LOGO = IconLoader.getIcon("/icons/moodle.svg", javaClass)
-}
