--- src/main/kotlin/il/co/sysbind/intellij/moodledev/MoodlePluginGeneratorSettings.kt
@@ -1,6 +1,0 @@
-package il.co.sysbind.intellij.moodledev
-
-data class MoodlePluginGeneratorSettings (
-    var pluginName: String = "",
-    var pluginType: String = ""
-)
