--- src/main/kotlin/il/co/sysbind/intellij/moodledev/actions/generation/MoodleJavaScriptFileAction.kt
@@ -1,56 +1,0 @@
-package il.co.sysbind.intellij.moodledev.actions.generation
-
-import com.intellij.ide.actions.CreateFileFromTemplateAction
-import com.intellij.ide.actions.CreateFileFromTemplateDialog
-import com.intellij.ide.fileTemplates.FileTemplate
-import com.intellij.ide.fileTemplates.FileTemplateManager
-import com.intellij.ide.fileTemplates.FileTemplateUtil
-import com.intellij.lang.javascript.JavaScriptFileType
-import com.intellij.openapi.actionSystem.CommonDataKeys
-import com.intellij.openapi.actionSystem.DataContext
-import com.intellij.openapi.project.DumbAware
-import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiDirectory
-import com.intellij.psi.PsiFile
-import il.co.sysbind.intellij.moodledev.project.MoodleProjectSettings
-import il.co.sysbind.intellij.moodledev.util.MoodleCorePathUtil
-
-
-class MoodleJavaScriptFileAction: CreateFileFromTemplateAction(CAPTION, "", JavaScriptFileType.INSTANCE.icon)
-    , DumbAware {
-
-    private companion object {
-        private const val CAPTION = "Moodle Javascript File"
-    }
-
-    override fun isAvailable(dataContext: DataContext?): Boolean {
-        if (!super.isAvailable(dataContext)) return false
-        val project = CommonDataKeys.PROJECT.getData(dataContext!!) ?: return false
-        CommonDataKeys.VIRTUAL_FILE.getData(dataContext) ?: return false
-        val moodle = project.getService(MoodleProjectSettings::class.java).settings
-        return moodle.pluginEnabled
-    }
-
-    override fun buildDialog(project: Project, directory: PsiDirectory, builder: CreateFileFromTemplateDialog.Builder) {
-        builder.setTitle(PhpCreateFileAction.CAPTION)
-            .addKind("Empty file", JavaScriptFileType.INSTANCE.icon, "Moodle JavaScript File")
-    }
-
-    override fun getActionName(directory: PsiDirectory?, newName: String, templateName: String?): String = CAPTION
-
-    override fun createFileFromTemplate(name: String, template: FileTemplate, dir: PsiDirectory): PsiFile {
-        val project = dir.project
-        val settings = project.getService(MoodleProjectSettings::class.java).settings
-
-        // Create properties for the template
-        val templateManager = FileTemplateManager.getInstance(project)
-        val props = templateManager.defaultProperties
-        props["USER_NAME"] = settings.userName
-        props["USER_EMAIL"] = settings.userEmail
-
-        // Extract PLUGIN_NAME from the closest version.php
-        props["MODULE_DIR"] = MoodleCorePathUtil.getModuleName(dir, "js")
-
-        return FileTemplateUtil.createFromTemplate(template, name, props, dir).containingFile
-    }
-}
