--- src/main/kotlin/il/co/sysbind/intellij/moodledev/project/MoodleSettings.kt
@@ -1,10 +1,0 @@
-package il.co.sysbind.intellij.moodledev.project
-
-class MoodleSettings {
-    var pluginEnabled: Boolean = false
-    var moodlePath: String = ""
-    var userName: String = ""
-    var userEmail: String = ""
-    var moodleVersion: String = "master"
-    val defaultMoodleTablePrefix = "mdl_"
-}
