+++ src/main/kotlin/il/co/sysbind/moodledev/MoodleBundle.kt
@@ -1,0 +1,21 @@
+package il.co.sysbind.moodledev
+
+import com.intellij.AbstractBundle
+import org.jetbrains.annotations.NonNls
+import org.jetbrains.annotations.PropertyKey
+
+@NonNls
+private const val BUNDLE = "messages.MoodleBundle"
+
+object MoodleBundle : AbstractBundle(BUNDLE) {
+
+    @Suppress("SpreadOperator")
+    @JvmStatic
+    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) = getMessage(key, *params)
+
+    @Suppress("SpreadOperator")
+    @JvmStatic
+    fun messagePointer(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) = run {
+        message(key, *params)
+    }
+}
