+++ src/main/kotlin/il/co/sysbind/moodledev/MoodleIcons.kt
@@ -1,0 +1,7 @@
+package il.co.sysbind.moodledev
+
+import com.intellij.openapi.util.IconLoader
+
+object MoodleIcons {
+    val LOGO = IconLoader.getIcon("/icons/moodle.svg", javaClass)
+}
