+++ src/main/kotlin/il/co/sysbind/moodledev/MoodlePluginGeneratorSettings.kt
@@ -1,0 +1,6 @@
+package il.co.sysbind.moodledev
+
+data class MoodlePluginGeneratorSettings (
+    var pluginName: String = "",
+    var pluginType: String = ""
+)
