+++ src/main/kotlin/il/co/sysbind/moodledev/listeners/MoodleManagerListener.kt
@@ -1,0 +1,13 @@
+package il.co.sysbind.moodledev.listeners
+
+
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.startup.StartupActivity
+import il.co.sysbind.moodledev.project.MoodleProjectSettings
+
+internal class MoodleManagerListener :  StartupActivity{
+
+    override fun runActivity(project: Project) {
+        project.getService(MoodleProjectSettings::class.java)
+    }
+}
