+++ src/main/kotlin/il/co/sysbind/moodledev/project/MoodleSettings.kt
@@ -1,0 +1,10 @@
+package il.co.sysbind.moodledev.project
+
+class MoodleSettings {
+    var pluginEnabled: Boolean = false
+    var moodlePath: String = ""
+    var userName: String = ""
+    var userEmail: String = ""
+    var moodleVersion: String = "master"
+    val defaultMoodleTablePrefix = "mdl_"
+}
