+++ src/main/kotlin/il/co/sysbind/moodledev/project/wizard/MoodlePlugin.kt
@@ -1,0 +1,10 @@
+package il.co.sysbind.moodledev.project.wizard
+
+import com.intellij.ide.util.projectWizard.ModuleBuilder
+import com.intellij.openapi.module.ModuleType
+
+class MoodlePlugin : ModuleBuilder() {
+    override fun getModuleType(): ModuleType<*> {
+        return ModuleType.EMPTY
+    }
+}
